dev_requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile dev_requirements.in
#
anyio==3.7.1
    # via
    #   -r dev_requirements.in
    #   httpcore
black==23.3.0
    # via -r dev_requirements.in
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   black
    #   flask
coverage[toml]==7.2.7
    # via
    #   -r dev_requirements.in
    #   pytest-cov
exceptiongroup==1.1.3
    # via
    #   anyio
    #   pytest
flake8==5.0.4
    # via -r dev_requirements.in
flask==2.2.5
    # via -r dev_requirements.in
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via -r dev_requirements.in
humanize==4.6.0
    # via -r dev_requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   yarl
importlib-metadata==4.2.0
    # via
    #   -r dev_requirements.in
    #   flask
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
multidict==6.0.4
    # via yarl
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   black
    #   pytest
pathspec==0.11.2
    # via black
platformdirs==3.11.0
    # via black
pluggy==1.2.0
    # via
    #   -r dev_requirements.in
    #   pytest
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pytest==7.4.2
    # via
    #   -r dev_requirements.in
    #   pytest-cov
pytest-cov==4.1.0
    # via -r dev_requirements.in
pyyaml==6.0.1
    # via vcrpy
six==1.16.0
    # via vcrpy
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   pytest
typing-extensions==4.7.1
    # via
    #   -r dev_requirements.in
    #   black
urllib3==1.26.17
    # via vcrpy
vcrpy==4.4.0
    # via -r dev_requirements.in
werkzeug==2.2.3
    # via
    #   -r dev_requirements.in
    #   flask
wrapt==1.15.0
    # via vcrpy
yarl==1.9.2
    # via vcrpy
zipp==3.15.0
    # via
    #   -r dev_requirements.in
    #   importlib-metadata
